+++ quoteAnime/Presentation/Components/SafariView.swift
@@ -1,0 +1,15 @@
+import SwiftUI
+import SafariServices
+
+struct SafariView: UIViewControllerRepresentable {
+    let url: URL
+
+    func makeUIViewController(context: Context) -> SFSafariViewController {
+        let vc = SFSafariViewController(url: url)
+        vc.preferredBarTintColor = UIColor(Color.surface)
+        vc.preferredControlTintColor = UIColor(Color.accentPurple)
+        return vc
+    }
+
+    func updateUIViewController(_ uiViewController: SFSafariViewController, context: Context) {}
+}